fix bottomSheet and filter

+++ app/src/main/java/com/lavyshyk/countrycity/ui/countryListFragment/MyFilter.kt
@@ -4,5 +4,5 @@ open class MyFilter(var lArea: Float,
                     var rArea: Float,
                     var lPopulation: Float,
                     var rPopulation: Float,
-                    var distance: Int,
+                    var distance: Float,
 )
